move class_features_data functions into another file
