+++ tests/test_class_features_data.py
@@ -1,6 +1,10 @@
 import pytest
 from dnd_character import Character
 from dnd_character.classes import CLASSES
+from dnd_character.features import (
+    get_class_features_data,
+    reset_class_features_data_counters,
+)
 
 
 @pytest.mark.parametrize(
@@ -48,6 +52,29 @@ from dnd_character.classes import CLASSES
         ),
     ),
 )
-def test_class_features_data_exist(input, expected_value):
+def test_class_features_data_exist_after_init(input, expected_value):
     c = Character(**input)
     assert c.class_features_data == expected_value
+
+
+def test_reset_class_features_data_bard_level_4():
+    c = Character(classs=CLASSES["bard"], level=4, charisma=10)
+    # inspiration count should be 4 in this situation
+    data = get_class_features_data(c)
+    data["available_inspiration_count"] = 1
+    # at level 4 a short rest should NOT restore inspiration count
+    data = reset_class_features_data_counters(character=c, data=data, short_rest=True, long_rest=False)
+    assert data["available_inspiration_count"] == 1
+    # a long rest should always restore inspiration count
+    data = reset_class_features_data_counters(character=c, data=data, short_rest=False, long_rest=True)
+    assert data["available_inspiration_count"] == 4
+
+
+def test_reset_class_features_data_bard_level_5():
+    c = Character(classs=CLASSES["bard"], level=5, charisma=10)
+    # inspiration count should be 5 in this situation
+    data = get_class_features_data(c)
+    data["available_inspiration_count"] = 1
+    # at level 5 a short rest should restore inspiration count
+    data = reset_class_features_data_counters(character=c, data=data, short_rest=True, long_rest=False)
+    assert data["available_inspiration_count"] == 5
